ci: use portable du flags, dont truncate to int32 in powershell
Flag -Wunsafe-buffer-usage is unusable as of clang19/20 for both
compiler and linter.

+++ ci/x86_64-windows.ps1
@@ -4,12 +4,15 @@ function CheckLastExitCode {
     return 0
 }
 
-[int] $MAX_SIZE_B = 2147483648 #2 GB = 2*(1024)**3 B
+$MAX_SIZE_B = 2147483648 #2 GB = 2*(1024)**3 B
 [string] $ZIG_CACHE_DIR = $(zig env | jq '. "global_cache_dir"')
-[int] $CHECK_SIZE_B = [int]$(gci $ZIG_CACHE_DIR | measure Length -s).Sum
-if ($CHECK_SIZE_B -gt $MAX_SIZE_B) {
-  Remove-Item -Recurse -Force -ErrorAction Ignore -Path $ZIG_CACHE_DIR
-  CheckLastExitCode
+if (Test-Path $ZIG_CACHE_DIR) {
+  $CHECK_SIZE_B = (gci $ZIG_CACHE_DIR | measure Length -s).Sum
+  if ($CHECK_SIZE_B -gt $MAX_SIZE_B) {
+    Write-Host "Removing global zig cache dir.."
+    Remove-Item -Recurse -Force -ErrorAction Ignore -Path $ZIG_CACHE_DIR
+    CheckLastExitCode
+  }
 }
 
 zig env
